% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strict-package.R
\docType{package}
\name{strict-package}
\alias{strict}
\alias{strict-package}
\title{strict: Make R Just a Little Stricter}
\description{
This packages tweaks the operation of base R code to make things a little stricter.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hadley/strict}
  \item \url{https://hadley.github.io/strict/}
  \item Report bugs at \url{https://github.com/hadley/strict/issues}
}

}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@posit.co}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
